refactor(services): extract web cli tool and config clients
Seventeen CLI methods move out across three bounded contexts: local CLI tool
detection and package operations, CLI parameter profiles, and CLI global
configuration profiles.

`memoryCliParameterSelections` and `webCliConfigProfiles` move with their
methods and stay unexported. The parameter selections are the second caller of
`web-mock-storage.ts`, keeping the same key and the same fallback-on-corruption
behavior; they hold catalog flag choices, not credentials.

`applyCliConfigProfile` stays inline. It asserts that switching a profile leaves
`workflowState` and the active session untouched — an invariant it can only check
by reading state this context does not own — so it reaches the profile and the
status through `findWebCliConfigProfile` and `setWebCliConfigStatus`.

The extracted CLI tool module calls `i18n.t` directly instead of copying the
root's `tr` wrapper, which stays behind for the ~40 call sites still in the root.
